so verifty token middle ware added and added to route3s

+++ server/routes/comments.js
@@ -3,11 +3,13 @@ const router = express.Router();
 const Post = require("../models/post");
 const Comment = require("../models/comment");
 const User = require("../models/user");
-
+const verifyToken = require("../middleware/auth");
 //add comment
 
-router.post("/add", async (req, res) => {
-  const { comment, userId, userName, postId } = req.body;
+router.post("/add", verifyToken, async (req, res) => {
+  const userId = req.userId;
+  const { comment, userName, postId } = req.body;
+
   try {
     const newComment = await Comment.create({
       comment,
@@ -23,16 +25,22 @@ router.post("/add", async (req, res) => {
 
 // delete comment
 
-router.delete("/delete/:id", async (req, res) => {
+router.delete("/delete/:id", verifyToken, async (req, res) => {
   const { id } = req.params;
+  const comment = await Comment.findById(id);
+  if (!comment) {
+    res.status(404).json({ message: "Comment not found" });
+  }
+  if (comment.userId.toString() !== req.userId) {
+    return res
+      .status(403)
+      .json({ message: "You can only delete your own comments" });
+  }
   try {
-    const deletedComment = await Comment.findByIdAndDelete(id);
-    if (!deletedComment) {
-      return res.status(404).json({ message: "Comment not found" });
-    }
+    await Comment.findByIdAndDelete(id);
     res
       .status(200)
-      .json({ message: "Comment deleted successfully", data: deletedComment });
+      .json({ message: "Comment deleted successfully", status: true });
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
@@ -54,13 +62,23 @@ router.get("/getByPost/:postId", async (req, res) => {
 
 // update comment
 
-router.put("/update/:id", async (req, res) => {
+router.put("/update/:id", verifyToken, async (req, res) => {
   const { id } = req.params;
+  if (req.userId !== id) {
+    return res.status(401).json({ message: "unauthorized", status: false });
+  }
+
   try {
     const comment = await Comment.findById(id);
     if (!comment) {
       return res.status(404).json({ message: "Comment not found" });
     }
+    if (comment.userId.toString() !== req.userId) {
+      return res
+        .status(403)
+        .json({ message: "You can only update your own comments" });
+    }
+
     const updatedComment = await Comment.findByIdAndUpdate(
       id,
       { $set: req.body },
